# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 経理・会計オールインワンAIアプリケーション API
  version: "1.0.0"
  description: |
    中小企業〜中堅企業向け「経理・会計をこのアプリ1つで完結させる」オールインワンAIプロダクトのAPI仕様。

    対応する統合DB設計は `docs/03_database_design.md` / `sql/001_initial_schema_all_in_one.sql` を参照。

    ## 設計原則
    - AIは提案(`/ai/suggestions`)のみを行い、確定処理(仕訳確定・税額計算等)は本APIの決定的ロジックが行う。
    - 状態遷移・業務アクションはサブリソース(例: `POST /journal-entries/{id}/post`)として明示する。
    - 全エンドポイントはマルチテナントであり、`X-Tenant-ID` ヘッダーとJWTクレームの双方で検証される。
    - 副作用を伴うPOST操作は `Idempotency-Key` ヘッダーによる冪等性をサポートする。
  contact:
    name: 経理会計オールインワンソフト開発チーム
servers:
  - url: https://api.keiri-kaikei.example.com/v1
    description: 本番環境
  - url: https://api.staging.keiri-kaikei.example.com/v1
    description: ステージング環境
  - url: http://localhost:3000/v1
    description: ローカル開発環境
...
security:
  - BearerAuth: []

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /expense-reports:
    get:
      tags: [ExpenseReports]
      summary: 経費精算申請を検索する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ExpenseReportStatus" }
        - name: on_behalf_of
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: 経費精算申請一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseReportListResponse" }
    post:
      tags: [ExpenseReports]
      summary: |
        経費精算を申請する(本人申請/代理申請)。
        申請と同時に先行draft仕訳が生成され、承認ルールが評価される。
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpenseReportCreate" }
      responses:
        "201":
          description: 申請成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseReportResponse" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /expense-reports/ocr:
    post:
      tags: [ExpenseReports]
      summary: |
        レシート・領収書画像をVision AI(設定済みのAIプロバイダー)で解析し、取引情報を
        `ai_suggestions`(target_type='expense_report', suggestion_type='ocr')へ追記保存する。
        1枚の画像内に複数のレシート・領収書が写っている場合は、検出した数だけ個別の
        提案レコードを追記し、その数だけ`suggestions`配列の要素として返す。
        `expense_report_id`は既存申請への明細追加時はその申請id、新規申請作成中は
        `POST /expense-reports`にも渡す予定のUUIDを指定する。
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, expense_report_id]
              properties:
                file:
                  type: string
                  format: binary
                  description: 領収書・レシート画像(JPEG/PNG/PDF)
                expense_report_id:
                  type: string
                  format: uuid
      responses:
        "201":
          description: 抽出成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseReportOcrResultResponse" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /expense-reports/{id}:
    get:
      tags: [ExpenseReports]
      summary: 経費精算申請詳細(明細含む)を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 経費精算申請詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseReportResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /expense-reports/{id}/lines:
    post:
      tags: [ExpenseReports]
      summary: 経費明細を追加する(交通費/出張旅費/日当等)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpenseReportLineCreate" }
      responses:
        "201":
          description: 追加成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseReportLineResponse" }

  # ----------------------------------------------------------------------------
  # Approval Requests (共通承認ワークフロー)
  # ----------------------------------------------------------------------------
  /approval-requests:
    get:
      tags: [ApprovalRequests]
      summary: 承認待ち・承認済みの申請を検索する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ApprovalRequestStatus" }
        - name: target_type
          in: query
          schema:
            type: string
            enum: [journal_entry, expense_report, vendor_bill]
        - name: pending_for_me
          in: query
          description: 現在のユーザーが次の承認ステップの承認者である申請のみに絞り込む
          schema: { type: boolean }
      responses:
        "200":
          description: 承認申請一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequestListResponse" }
  /approval-requests/{id}:
    get:
      tags: [ApprovalRequests]
      summary: 承認申請詳細(承認履歴含む)を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 承認申請詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequestResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /approval-requests/{id}/approve:
    post:
      tags: [ApprovalRequests]
      summary: 承認する
      description: |
        申請者本人による承認(自己承認)は `SELF_APPROVAL_NOT_ALLOWED` エラーで
        409を返す(DB側 `fn_prevent_self_approval` トリガーによる最終防御)。
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
      responses:
        "200":
          description: 承認成功(全ステップ完了時は対象リソースがposted/approvedへ遷移)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequestResponse" }
        "409":
          description: 自己承認、または既に完了した申請への操作
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                self_approval:
                  value:
                    success: false
                    error:
                      code: SELF_APPROVAL_NOT_ALLOWED
                      message: "申請者本人は承認者になれません"
                      details: []
  /approval-requests/{id}/reject:
    post:
      tags: [ApprovalRequests]
      summary: 却下する(対象リソースはdraft等へ差し戻される)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment: { type: string }
      responses:
        "200":
          description: 却下成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApprovalRequestResponse" }

  # ----------------------------------------------------------------------------
  # Recurring Journal Templates (定期仕訳)
  # ----------------------------------------------------------------------------
...
  /fixed-assets:
    get:
      tags: [FixedAssets]
      summary: 固定資産台帳を検索する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/FixedAssetStatus" }
      responses:
        "200":
          description: 固定資産一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FixedAssetListResponse" }
    post:
      tags: [FixedAssets]
      summary: 固定資産を登録する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FixedAssetCreate" }
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FixedAssetResponse" }
  /fixed-assets/{id}:
    get:
      tags: [FixedAssets]
      summary: 固定資産詳細(償却スケジュール含む)を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 固定資産詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FixedAssetResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [FixedAssets]
      summary: 固定資産情報を更新する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FixedAssetCreate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FixedAssetResponse" }
  /fixed-assets/{id}/dispose:
    post:
      tags: [FixedAssets]
      summary: 固定資産を除却・売却する(除却損/売却損益を自動仕訳)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [disposal_date, disposal_type]
              properties:
                disposal_date: { type: string, format: date }
                disposal_type: { type: string, enum: [disposed, sold] }
                proceeds_amount: { type: number, default: 0 }
      responses:
        "200":
          description: 除却・売却成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FixedAssetResponse" }
  /fixed-assets/depreciation-runs:
    post:
      tags: [FixedAssets]
      summary: 月次減価償却バッチを実行する(稼働中の全資産を対象にdraft仕訳を一括生成)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fiscal_period_id]
              properties:
                fiscal_period_id: { type: string, format: uuid }
                asset_ids:
                  type: array
                  description: 未指定時は稼働中の全資産が対象
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: 償却バッチ実行結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { const: true }
                  data:
                    type: object
                    properties:
                      processed_count: { type: integer }
                      journal_entry_ids:
                        type: array
                        items: { type: string, format: uuid }

  # ----------------------------------------------------------------------------
  # Reports (試算表・財務三表)
  # ----------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWTアクセストークン。`tenant_id` クレームを含み、`X-Tenant-ID` ヘッダーと
        一致しない場合はAPI層で403 (`TENANT_MISMATCH`) を返す。
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: 操作対象テナントのUUID。JWTクレームの tenant_id と一致する必要がある。
      schema: { type: string, format: uuid }
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        副作用を伴うPOST操作の冪等性を保証するクライアント生成キー(推奨: UUID)。
        同一キーでの再送は最初のレスポンスをそのまま返す。
      schema: { type: string }
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
...
    NotFound:
      description: リソースが見つかりません(RLSにより越境アクセス時も同一の404を返す)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    UnprocessableEntity:
      description: 入力値のバリデーションエラー
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    Meta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        pagination:
          type: object
          properties:
            page: { type: integer }
            page_size: { type: integer }
            total_count: { type: integer }
            total_pages: { type: integer }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            例: INVALID_BALANCE, VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED, FORBIDDEN,
            TENANT_MISMATCH, IDEMPOTENCY_KEY_CONFLICT, VOID_WINDOW_EXPIRED,
            SELF_APPROVAL_NOT_ALLOWED, DUPLICATE_IMPORT, CONFLICT, ALREADY_EXISTS,
            INVITATION_EXPIRED, INVITATION_REVOKED, INVITATION_ALREADY_ACCEPTED
        message: { type: string }
        details:
          type: array
          items: { type: object }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: "#/components/schemas/ApiError" }

    # ---- Enums (DBのENUM型と対応) ----
...
    ExpenseReportStatus:
      type: string
      enum: [submitted, in_review, approved, rejected, reimbursement_scheduled, reimbursed]
    PaymentMethod:
      type: string
      enum: [cash, corporate_card, bank_transfer, employee_advance]
    ApprovalAction:
      type: string
      enum: [approve, reject]
    ApprovalRequestStatus:
      type: string
      enum: [pending, approved, rejected]
    DepreciationMethod:
      type: string
      enum: [straight_line, declining_balance]
    FixedAssetStatus:
      type: string
      enum: [active, disposed, sold]
...
    ExpenseReportLine:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        line_no: { type: integer }
        expense_date: { type: string, format: date }
        category_id: { type: string, format: uuid }
        description: { type: string, nullable: true }
        amount: { type: number, exclusiveMinimum: 0 }
        payment_method: { $ref: "#/components/schemas/PaymentMethod" }
        tax_category_id: { type: string, format: uuid, nullable: true }
        card_transaction_id: { type: string, format: uuid, nullable: true }
    ExpenseReportLineCreate:
      type: object
      required: [expense_date, category_id, amount, payment_method]
      properties:
        expense_date: { type: string, format: date }
        category_id: { type: string, format: uuid }
        description: { type: string }
        amount: { type: number, exclusiveMinimum: 0 }
        payment_method: { $ref: "#/components/schemas/PaymentMethod" }
        tax_category_id: { type: string, format: uuid }
        card_transaction_id: { type: string, format: uuid }
    ExpenseReportLineResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/ExpenseReportLine" }
        meta: { $ref: "#/components/schemas/Meta" }
    ExpenseReport:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        report_no: { type: string, readOnly: true }
        submitted_by: { type: string, format: uuid, readOnly: true }
        on_behalf_of: { type: string, format: uuid }
        purpose: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ExpenseReportStatus" }
        total_amount: { type: number, readOnly: true }
        journal_entry_id: { type: string, format: uuid, nullable: true, readOnly: true }
        lines:
          type: array
          items: { $ref: "#/components/schemas/ExpenseReportLine" }
    ExpenseReportCreate:
      type: object
      required: [on_behalf_of, lines]
      properties:
        id:
          type: string
          format: uuid
          description: |
            省略可。レシートOCR機能でフロントエンドが撮影段階で事前生成したUUIDを指定すると、
            そのUUIDを対象にした`ai_suggestions`(target_type='expense_report')のOCR提案の
            target_idと、確定後の申請idが一致する。省略時はサーバー側で自動採番する。
        on_behalf_of:
          type: string
          format: uuid
          description: 代理申請の場合は費用が帰属する従業員のID。本人申請時は自身のID。
        purpose: { type: string }
        lines:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ExpenseReportLineCreate" }
    ExpenseReportOcrSuggestionData:
      type: object
      readOnly: true
      required: [confidence_score, model_name]
      properties:
        transaction_date: { type: string, format: date, nullable: true }
        amount: { type: number, nullable: true }
        vendor_name: { type: string, nullable: true }
        invoice_registration_number: { type: string, nullable: true }
        suggested_account_id: { type: string, format: uuid, nullable: true }
        confidence_score: { type: number, minimum: 0, maximum: 1 }
        model_name: { type: string }
    ExpenseReportOcrSuggestion:
      type: object
      readOnly: true
      required: [id, data]
      properties:
        id:
          type: string
          format: uuid
          description: 追記された `ai_suggestions` レコードのid
        data: { $ref: "#/components/schemas/ExpenseReportOcrSuggestionData" }
    ExpenseReportOcrResult:
      type: object
      readOnly: true
      required: [suggestions]
      properties:
        suggestions:
          type: array
          description: 画像内から検出されたレシート数ぶんの提案(1枚に複数のレシートが写っている場合は複数件)
          items: { $ref: "#/components/schemas/ExpenseReportOcrSuggestion" }
    ExpenseReportOcrResultResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/ExpenseReportOcrResult" }
        meta: { $ref: "#/components/schemas/Meta" }
    ExpenseReportResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/ExpenseReport" }
        meta: { $ref: "#/components/schemas/Meta" }
    ExpenseReportListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/ExpenseReport" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Approval Requests ----
    ApprovalHistoryEntry:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        step_number: { type: integer }
        approver_id: { type: string, format: uuid }
        action: { $ref: "#/components/schemas/ApprovalAction" }
        comment: { type: string, nullable: true }
        acted_at: { type: string, format: date-time }
    ApprovalRequest:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        target_type: { type: string, enum: [journal_entry, expense_report, vendor_bill] }
        target_id: { type: string, format: uuid }
        submitted_by: { type: string, format: uuid }
        total_steps: { type: integer }
        current_step: { type: integer }
        status: { $ref: "#/components/schemas/ApprovalRequestStatus" }
        history:
          type: array
          items: { $ref: "#/components/schemas/ApprovalHistoryEntry" }
    ApprovalRequestResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/ApprovalRequest" }
        meta: { $ref: "#/components/schemas/Meta" }
    ApprovalRequestListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/ApprovalRequest" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Recurring Journal Templates ----
...
    DepreciationSchedule:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        fiscal_period_id: { type: string, format: uuid }
        scheduled_amount: { type: number }
        actual_amount: { type: number, nullable: true }
        journal_entry_id: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [scheduled, posted] }
    FixedAsset:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        asset_no: { type: string, readOnly: true }
        name: { type: string }
        category: { type: string, nullable: true }
        acquisition_date: { type: string, format: date }
        acquisition_cost: { type: number, exclusiveMinimum: 0 }
        useful_life_years: { type: integer, exclusiveMinimum: 0 }
        depreciation_method: { $ref: "#/components/schemas/DepreciationMethod" }
        salvage_value: { type: number, default: 0 }
        status: { $ref: "#/components/schemas/FixedAssetStatus" }
        accumulated_depreciation: { type: number, readOnly: true }
        department_id: { type: string, format: uuid, nullable: true }
        asset_account_id: { type: string, format: uuid }
        depreciation_expense_account_id: { type: string, format: uuid }
        depreciation_schedules:
          type: array
          items: { $ref: "#/components/schemas/DepreciationSchedule" }
    FixedAssetCreate:
      type: object
      required:
        - name
        - acquisition_date
        - acquisition_cost
        - useful_life_years
        - depreciation_method
        - asset_account_id
        - depreciation_expense_account_id
      properties:
        name: { type: string }
        category: { type: string }
        acquisition_date: { type: string, format: date }
        acquisition_cost: { type: number, exclusiveMinimum: 0 }
        useful_life_years: { type: integer, exclusiveMinimum: 0 }
        depreciation_method: { $ref: "#/components/schemas/DepreciationMethod" }
        salvage_value: { type: number, default: 0 }
        department_id: { type: string, format: uuid }
        asset_account_id: { type: string, format: uuid }
        depreciation_expense_account_id: { type: string, format: uuid }
    FixedAssetResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/FixedAsset" }
        meta: { $ref: "#/components/schemas/Meta" }
    FixedAssetListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/FixedAsset" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Reports ----
...